User_management/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
